Splitting main loop in a fetcher-and-worker model
- Fetchers are responsible for pulling potential jobs to work on from the persistent queue in the DB, and place them in an in-memory channel;
- Workers take jobs from the in-memory channel, attempt to lock them, and if successful, can proceed to process them.

The idea is that workers are fed what to work on, while other pieces can be responsible for deciding what jobs should be attemped. We could imagine placing jobs in the queue using other triggers than simply the DB, such as a post-commit hook on enqueuing which would speed up the discovery of a job and shorten the enqueue-to-completion timeframe.